claims metadata preloading was added
